% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_demographics.R
\name{eiof_check_gdqs}
\alias{eiof_check_gdqs}
\title{Check the Global Diet Quality Score variable}
\usage{
eiof_check_gdqs(gdqs, cutoff = 23, max_score = 49)
}
\arguments{
\item{gdqs}{Numeric vector of GDQS values.}

\item{cutoff}{The healthy-diet threshold, default 23.}

\item{max_score}{The maximum attainable score, used for the range check.
Default 49.}
}
\value{
An object of class `eiof_check`.
}
\description{
Check the Global Diet Quality Score variable
}
\examples{
data(eiof_demo)
eiof_check_gdqs(eiof_demo$gdqs)

}
